+++ src/VaultSharp/V1/SecretsEngines/Transform/DecodedResponse.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.Transform
+{
+    /// <summary>
+    /// Response for decoding.
+    /// </summary>
+    public class DecodedResponse : DecodedItem
+    {
+        /// <summary>
+        /// Decoded items.
+        /// </summary>
+        [JsonProperty("batch_results")]
+        public List<DecodedItem> DecodedItems { get; set; }
+    }
+}