+++ src/VaultSharp/V1/SecretsEngines/Transform/EncodedItem.cs
@@ -1,0 +1,22 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.Transform
+{
+    /// <summary>
+    /// Represents a single Encoded item.
+    /// </summary>
+    public class EncodedItem
+    {
+        /// <summary>
+        /// Specifies the encoded value.
+        /// </summary>
+        [JsonProperty("encoded_value")]
+        public string EncodedValue { get; set; }
+
+        /// <summary>
+        /// Specifies the base64 encoded tweak that was provided during encoding.
+        /// </summary>
+        [JsonProperty("tweak")]
+        public string Tweak { get; set; }
+    }
+}