+++ src/VaultSharp/V1/SecretsEngines/Transform/EncodedResponse.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.Transform
+{
+    /// <summary>
+    /// Response for encoding.
+    /// </summary>
+    public class EncodedResponse : EncodedItem
+    {
+        /// <summary>
+        /// Encoded items.
+        /// </summary>
+        [JsonProperty("batch_results")]
+        public List<EncodedItem> EncodedItems { get; set; }
+    }
+}